+++ src/main/java/com/feelory/feelory_backend/words/controller/CategoriesController.java
@@ -3,11 +3,17 @@ package com.feelory.feelory_backend.words.controller;
 import com.feelory.feelory_backend.global.api.ApiResponse;
 import com.feelory.feelory_backend.global.api.SuccessCode;
 import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
-import com.feelory.feelory_backend.words.model.*;
+import com.feelory.feelory_backend.words.dto.request.CategoryCreateRequest;
+import com.feelory.feelory_backend.words.dto.request.CategoryDeleteRequest;
+import com.feelory.feelory_backend.words.dto.request.CategoryListRequest;
+import com.feelory.feelory_backend.words.dto.request.CategoryUpdateRequest;
+import com.feelory.feelory_backend.words.dto.response.CategoryCreateResponse;
+import com.feelory.feelory_backend.words.dto.response.CategoryDeleteResponse;
+import com.feelory.feelory_backend.words.dto.response.CategoryListResponse;
+import com.feelory.feelory_backend.words.dto.response.CategoryUpdateResponse;
 import com.feelory.feelory_backend.words.service.CategoriesService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springdoc.core.annotations.ParameterObject;
