+++ src/main/java/com/feelory/feelory_backend/words/controller/DailyWordsController.java
@@ -5,7 +5,13 @@ import com.feelory.feelory_backend.global.api.SuccessCode;
 import com.feelory.feelory_backend.global.exception.exceptions.common.InvalidDateFormatException;
 import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.words.docs.DailyWordsDocs;
-import com.feelory.feelory_backend.words.model.*;
+import com.feelory.feelory_backend.words.dto.request.DailyWordCreateRequest;
+import com.feelory.feelory_backend.words.dto.request.DailyWordDeleteRequest;
+import com.feelory.feelory_backend.words.dto.request.DailyWordUpdateRequest;
+import com.feelory.feelory_backend.words.dto.response.DailyWordCreateResponse;
+import com.feelory.feelory_backend.words.dto.response.DailyWordDeleteResponse;
+import com.feelory.feelory_backend.words.dto.response.DailyWordDetailResponse;
+import com.feelory.feelory_backend.words.dto.response.DailyWordUpdateResponse;
 import com.feelory.feelory_backend.words.service.DailyWordsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
