+++ src/main/java/com/feelory/feelory_backend/words/controller/WordsController.java
@@ -4,7 +4,14 @@ package com.feelory.feelory_backend.words.controller;
 import com.feelory.feelory_backend.global.api.ApiResponse;
 import com.feelory.feelory_backend.global.api.SuccessCode;
 import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
-import com.feelory.feelory_backend.words.model.*;
+import com.feelory.feelory_backend.words.dto.request.WordCreateRequest;
+import com.feelory.feelory_backend.words.dto.request.WordDeleteRequest;
+import com.feelory.feelory_backend.words.dto.request.WordListRequest;
+import com.feelory.feelory_backend.words.dto.request.WordUpdateRequest;
+import com.feelory.feelory_backend.words.dto.response.WordCreateResponse;
+import com.feelory.feelory_backend.words.dto.response.WordDeleteResponse;
+import com.feelory.feelory_backend.words.dto.response.WordListResponse;
+import com.feelory.feelory_backend.words.dto.response.WordUpdateResponse;
 import com.feelory.feelory_backend.words.service.WordsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
