+++ src/main/java/com/feelory/feelory_backend/words/dto/request/CategoryListRequest.java
@@ -1,4 +1,4 @@
-package com.feelory.feelory_backend.words.model;
+package com.feelory.feelory_backend.words.dto.request;
 
 import lombok.*;
