+++ src/main/java/com/feelory/feelory_backend/words/dto/response/DailyWordDetailResponse.java
@@ -1,5 +1,6 @@
-package com.feelory.feelory_backend.words.model;
+package com.feelory.feelory_backend.words.dto.response;
 
+import com.feelory.feelory_backend.words.dto.model.DailyWordDto;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -9,7 +10,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class DailyWordCreateResponse {
-    private Boolean isAlreadyAssigned;
+public class DailyWordDetailResponse {
     private DailyWordDto dailyWord;
 }
