+++ src/main/java/com/feelory/feelory_backend/words/service/CategoriesService.java
@@ -2,10 +2,17 @@ package com.feelory.feelory_backend.words.service;
 
 import com.feelory.feelory_backend.global.exception.exceptions.words.CategoryNotFoundException;
 import com.feelory.feelory_backend.global.exception.exceptions.words.DuplicateCategoryNameException;
-import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.global.util.ValidationUtil;
+import com.feelory.feelory_backend.words.dto.model.CategoryDto;
+import com.feelory.feelory_backend.words.dto.request.CategoryCreateRequest;
+import com.feelory.feelory_backend.words.dto.request.CategoryDeleteRequest;
+import com.feelory.feelory_backend.words.dto.request.CategoryListRequest;
+import com.feelory.feelory_backend.words.dto.request.CategoryUpdateRequest;
+import com.feelory.feelory_backend.words.dto.response.CategoryCreateResponse;
+import com.feelory.feelory_backend.words.dto.response.CategoryDeleteResponse;
+import com.feelory.feelory_backend.words.dto.response.CategoryListResponse;
+import com.feelory.feelory_backend.words.dto.response.CategoryUpdateResponse;
 import com.feelory.feelory_backend.words.entity.WordCategories;
-import com.feelory.feelory_backend.words.model.*;
 import com.feelory.feelory_backend.words.repository.CategoriesRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
