+++ src/main/java/com/feelory/feelory_backend/words/service/WordsService.java
@@ -3,11 +3,18 @@ package com.feelory.feelory_backend.words.service;
 import com.feelory.feelory_backend.global.exception.exceptions.words.CategoryNotFoundException;
 import com.feelory.feelory_backend.global.exception.exceptions.words.DuplicateWordNameException;
 import com.feelory.feelory_backend.global.exception.exceptions.words.WordNotFoundException;
-import com.feelory.feelory_backend.global.security.auth.jwt.JwtTokenProvider;
 import com.feelory.feelory_backend.global.util.ValidationUtil;
+import com.feelory.feelory_backend.words.dto.model.WordDto;
+import com.feelory.feelory_backend.words.dto.request.WordCreateRequest;
+import com.feelory.feelory_backend.words.dto.request.WordDeleteRequest;
+import com.feelory.feelory_backend.words.dto.request.WordListRequest;
+import com.feelory.feelory_backend.words.dto.request.WordUpdateRequest;
+import com.feelory.feelory_backend.words.dto.response.WordCreateResponse;
+import com.feelory.feelory_backend.words.dto.response.WordDeleteResponse;
+import com.feelory.feelory_backend.words.dto.response.WordListResponse;
+import com.feelory.feelory_backend.words.dto.response.WordUpdateResponse;
 import com.feelory.feelory_backend.words.entity.WordCategories;
 import com.feelory.feelory_backend.words.entity.Words;
-import com.feelory.feelory_backend.words.model.*;
 import com.feelory.feelory_backend.words.repository.CategoriesRepository;
 import com.feelory.feelory_backend.words.repository.WordsRepository;
 import lombok.RequiredArgsConstructor;
